website/requirements.txt
and==66.0.2
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
dj-database-url==2.0.0
Django==4.1.7
django-browser-reload==1.7.0
django-heroku==0.3.1
django-robots==5.0
django-tailwind==3.5.0
django-tinymce==3.6.1
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
psycopg2==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
